refactor function into a separate file
