+++ scheduler/pkg/server/server.go
@@ -594,20 +594,3 @@ func (s *SchedulerServer) SchedulerStatus(ctx context.Context, req *pb.Scheduler
 	}, nil
 }
 
-func sentWithTimeout(f func() error, d time.Duration) (bool, error) {
-	errChan := make(chan error, 1)
-	go func() {
-		errChan <- f()
-		close(errChan)
-	}()
-	t := time.NewTimer(d)
-	select {
-	case <-t.C:
-		return true, status.Errorf(codes.DeadlineExceeded, "Failed to send event within timeout")
-	case err := <-errChan:
-		if !t.Stop() {
-			<-t.C
-		}
-		return false, err
-	}
-}
